% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_terra_tiles.R
\name{create_tile_exts}
\alias{create_tile_exts}
\title{Create extents for raster tiles}
\usage{
create_tile_exts(raster, ncol, nrow)
}
\arguments{
\item{raster}{a SpatRaster object}

\item{ncol}{integer; number of columns to split the SpatRaster into}

\item{nrow}{integer; number of rows to split the SpatRaster into}
}
\value{
list of named numeric vectors with xmin, xmax, ymin, and ymax values
}
\description{
A wrapper around \code{\link[terra:makeTiles]{terra::getTileExtents()}} for creating a tile specification.
Rather than having to supply a template, one is created from the original
raster extent and CRS using \code{ncol} and \code{nrow}.  The output is a list of named
numeric vectors that can be coerced to SpatExtent objects with
\code{\link[terra:ext]{terra::ext()}}.
}
\note{
While this may have general use, it was created primarily for use
within \code{\link[=tar_terra_tiles]{tar_terra_tiles()}}.
}
\examples{
f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
create_tile_exts(r, ncol = 2, nrow = 2)
}
\author{
Eric Scott
}
