% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotargets-cache.R
\name{geotargets_destroy_cache}
\alias{geotargets_destroy_cache}
\alias{geotargets_init_cache}
\title{Manage geotargets Cache Files}
\usage{
geotargets_destroy_cache(name = NULL, init = FALSE)

geotargets_init_cache(name = NULL)
}
\arguments{
\item{name}{character. Target name (default: \code{NULL} will delete all target cache files)}

\item{init}{logical. Re-create empty cache directory? Default: \code{FALSE}}
}
\value{
integer. 0 for success, 1 for failure, invisibly.
}
\description{
The \code{geotargets} cache is a folder containing one subfolder for each target created using \code{tar_terra_rast_wrap()} method. Each subfolder contains one or more files containing data and metadata to support a \code{terra} \code{SpatRaster} object.
}
\details{
The cache directory can be customized using \code{geotargets_option_set("cache.dir")} or environment variable \code{GEOTARGETS_CACHE_DIR}.

Periodically you may want to purge files in the cache using \code{geotargets_destroy_cache()}.
}
\examples{
\dontrun{

# delete cache folder for target named "foo"
geotargets_destroy_cache("foo")

# create empty folder
geotargets_init_cache("foo")

# delete and recreate folder
geotargets_destroy_cache("foo", init = TRUE)

# delete all cache files
geotargets_destroy_cache()

}
}
