% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotargets-option.R
\name{geotargets_option_get}
\alias{geotargets_option_get}
\alias{geotargets_option_set}
\title{Get or Set geotargets Options}
\usage{
geotargets_option_get(x)

geotargets_option_set(x, value)
}
\arguments{
\item{x}{Character. Option name. See Details.}

\item{value}{Value to assign to option \code{x}.}
}
\description{
Get or set behavior for geospatial data target stores using geotargets-specific global options.
}
\details{
\subsection{Available Options}{
\itemize{
\item \code{"geotargets.raster.gdal_creation_options"} - set the GDAL creation options used when writing raster files to target store (default: \code{"ENCODING=UTF-8"})
\item \code{"geotargets.raster.gdal_driver_name"} - set the file type used for raster data in target store (default: \code{"GTiff"})
}

Each option can be overridden with a system environment variable. Options include:
\itemize{
\item \code{GEOTARGETS_RASTER_GDAL_CREATION_OPTIONS}
\item \code{GEOTARGETS_RASTER_GDAL_DRIVER_NAME}
}
}
}
