% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotargets-package.R
\docType{package}
\name{geotargets-package}
\alias{geotargets}
\alias{geotargets-package}
\title{geotargets: Targets extensions for geospatial formats}
\description{
Provides extensions for various geospatial file formats, such as shapefiles and rasters. See the vignettes for worked examples and demonstrations and explanations of how to use the various package extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njtierney/geotargets}
  \item \url{https://njtierney.github.io/geotargets/}
  \item Report bugs at \url{https://github.com/njtierney/geotargets/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Eric Scott (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
  \item Andrew Brown (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})
}

}
\keyword{internal}
