% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_terra_tiles.R
\name{set_window}
\alias{set_window}
\title{Copy a raster with a window}
\usage{
set_window(raster, window)
}
\arguments{
\item{raster}{A SpatRaster object}

\item{window}{A SpatExtent object}
}
\description{
Create a new SpatRaster object based on an original SpatRaster and a window.
This is a wrapper around \code{\link[terra:window]{terra::window()}} which, rather than modifying the
SpatRaster in place, returns a new SpatRaster leaving the original unchanged.
}
\note{
While this may have general use, it was created primarily for use
within \code{\link[=tar_terra_tiles]{tar_terra_tiles()}}.
}
\examples{
f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
e <- ext(c(5.9, 6,49.95, 50))
r2 <- set_window(rast, e)
ext(r)
ext(r2)

}
\author{
Eric Scott
}
