% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotargets-cache.R
\name{geotargets_destroy_cache}
\alias{geotargets_destroy_cache}
\alias{geotargets_init_cache}
\title{Manage geotargets Cache Files}
\usage{
geotargets_destroy_cache(
  cache_dir = geotargets_option_get("cache.dir"),
  init = FALSE
)

geotargets_init_cache(cache_dir = geotargets_option_get("cache.dir"))
}
\arguments{
\item{cache_dir}{character. Path to a directory to use as a cache}

\item{init}{logical. Re-create empty cache directory? Default: \code{FALSE}}
}
\value{
integer. 0 for success, 1 for failure, invisibly.
}
\description{
The \code{geotargets} cache is a folder created when the \code{use_cache} option is TRUE (see \code{\link[=geotargets_option_set]{geotargets_option_set()}}). The cache directory can be customized using \code{geotargets_option_set(cache_dir = )} or environment variable \code{GEOTARGETS_CACHE_DIR}.
}
\details{
Periodically you may want to purge files in the cache using \code{geotargets_destroy_cache()}.
}
\examples{
\dontrun{

# delete all cache files
geotargets_destroy_cache()

}
}
