% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotargets-package.R
\docType{package}
\name{geotargets-package}
\alias{geotargets}
\alias{geotargets-package}
\title{geotargets: 'targets' Extensions for Geographic Spatial Formats}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides extensions for various geographic spatial file formats, such as shape files and rasters. Currently provides support for the 'terra' geographic spatial formats. See the vignettes for worked examples, demonstrations, and explanations of how to use the various package extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/geotargets}
  \item \url{https://docs.ropensci.org/geotargets/}
  \item Report bugs at \url{https://github.com/ropensci/geotargets/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Eric Scott (\href{https://orcid.org/0000-0002-7430-7879}{ORCID})
  \item Andrew Brown (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})
}

}
\keyword{internal}
