% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_terra_tiles.R
\name{tile_grid}
\alias{tile_grid}
\alias{tile_blocksize}
\title{Helper functions to create tiles}
\usage{
tile_grid(raster, ncol, nrow)

tile_blocksize(raster)
}
\arguments{
\item{raster}{a SpatRaster object}

\item{ncol}{integer; number of columns to split the SpatRaster into}

\item{nrow}{integer; number of rows to split the SpatRaster into}
}
\value{
list of named numeric vectors with xmin, xmax, ymin, and ymax values
that can be coerced to SpatExtent objects with \code{\link[terra:ext]{terra::ext()}}.
}
\description{
Wrappers around \code{\link[terra:makeTiles]{terra::getTileExtents()}} that return a list of named numeric
vectors describing the extents of tiles rather than \code{SpatExtent} objects.
While these may have general use, they are intended primarily for supplying
to the \code{tile_fun} argument of \code{\link[=tar_terra_tiles]{tar_terra_tiles()}}.
}
\details{
\code{tile_blocksize()} creates extents using the raster's native blocksize (see
\code{\link[terra:readwrite]{terra::fileBlocksize()}}), which should be more memory efficient. \code{tile_grid()}
allows specification of a number of rows and columns to split the raster
into.  E.g. nrow = 2 and ncol = 2 would create 4 tiles (because it specifies a 2x2 matrix, which has 4 elements).
}
\examples{
f <- system.file("ex/elev.tif", package="terra")
r <- terra::rast(f)
r_tiles <- tile_grid(r, ncol = 2, nrow = 2)
r_tiles
}
\author{
Eric Scott
}
